ok, first commit from my latest (and probably not last) Linux installation, which means I had to get a new SSH key and everything. Still, I'm on ubuntu 11.10 now, and I'm learning to love Unity. i used to use XFCE or that weird Gnome hodge-podge that Mint came up with, and i'll say, Unity is pretty weird and somewhat buggy.
Back to the point - still working on fAde (source code title fightergame), and I must say, I like how it's turning out. Going to put in what's probably
the last feature (a timer showing how long it's been going) and then maybe move on to another project. May come back to fAde later because it's certainly
a neat concept that I want to elaborate on. Anyways.